rename block_header_state.{hpp,cpp} to block_header_state_legacy.{hpp, cpp} and rename block_state.{hpp,cpp} to block_state_legacy.{hpp,cpp}
