+++ libraries/chain/block_header_state_legacy.cpp
@@ -1,4 +1,4 @@
-#include <eosio/chain/block_header_state.hpp>
+#include <eosio/chain/block_header_state_legacy.hpp>
 #include <eosio/chain/exceptions.hpp>
 #include <limits>
