+++ libraries/chain/block_state_legacy.cpp
@@ -1,4 +1,4 @@
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/exceptions.hpp>
 
 namespace eosio { namespace chain {
