+++ libraries/chain/include/eosio/chain/controller.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/block_log.hpp>
 #include <eosio/chain/trace.hpp>
 #include <eosio/chain/genesis_state.hpp>
