+++ libraries/chain/include/eosio/chain/snapshot_scheduler.hpp
@@ -2,7 +2,7 @@
 
 #include <eosio/chain/pending_snapshot.hpp>
 
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/config.hpp>
 #include <eosio/chain/exceptions.hpp>
 #include <eosio/chain/resource_limits.hpp>
