+++ libraries/chain/include/eosio/chain/subjective_billing.hpp
@@ -2,7 +2,7 @@
 
 #include <eosio/chain/types.hpp>
 #include <eosio/chain/exceptions.hpp>
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/transaction.hpp>
 #include <eosio/chain/resource_limits.hpp>
 #include <eosio/chain/resource_limits_private.hpp>
