+++ plugins/chain_plugin/include/eosio/chain_plugin/trx_retry_db.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include <eosio/chain/types.hpp>
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain/trace.hpp>
 
 namespace eosio::chain_apis {
