+++ plugins/chain_plugin/test/test_account_query_db.cpp
@@ -2,7 +2,7 @@
 #include <eosio/chain/permission_object.hpp>
 #include <eosio/testing/tester.hpp>
 #include <eosio/chain/types.hpp>
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 #include <eosio/chain_plugin/account_query_db.hpp>
 #include <eosio/chain/thread_utils.hpp>
