+++ plugins/trace_api_plugin/test/test_extraction.cpp
@@ -5,7 +5,7 @@
 #include <eosio/chain/trace.hpp>
 #include <eosio/chain/transaction.hpp>
 #include <eosio/chain/block.hpp>
-#include <eosio/chain/block_state.hpp>
+#include <eosio/chain/block_state_legacy.hpp>
 
 #include <eosio/trace_api/test_common.hpp>
 #include <eosio/trace_api/chain_extraction.hpp>
